fix: removed guard from passion and moved back guard in users
