+++ passion-meet-api/src/passion/passion.module.ts
@@ -3,8 +3,6 @@ import { PassionService } from './passion.service';
 import { PassionController } from './passion.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Passion } from './passion.entity';
-import { APP_GUARD } from '@nestjs/core';
-import { AuthGuard } from '../auth.guard';
 
 @Module({
   imports: [
@@ -13,10 +11,6 @@ import { AuthGuard } from '../auth.guard';
   controllers: [PassionController],
   providers: [
     PassionService,
-    {
-      provide: APP_GUARD,
-      useClass: AuthGuard,
-    },
   ],
 })
 export class PassionModule {}
