+++ passion-meet-api/src/user/user.module.ts
@@ -3,7 +3,7 @@ import { UserService } from './user.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './user.entity';
 import { APP_GUARD } from '@nestjs/core';
-import { AuthGuard } from '../auth.guard';
+import { AuthGuard } from './auth.guard';
 import { UserController } from './user.controller';
 
 @Module({
